+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Changelog](http://keepachangelog.com/).
 
 
+
 ## Unreleased
+---
+
+### New
+
+### Changes
+
+### Fixes
+
+### Breaks
+
+
+## 1.1.0 - (2019-08-08)
 ---
 
 ### New
@@ -12,10 +25,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 * Only default to add available integrity if asset is from the manifest
-
-### Fixes
-
-### Breaks
 
 
 ## 1.0.0 - (2019-08-07)